# Optimizations to egraph framework